+++ libvmaf/src/feature/motion.h
@@ -16,4 +16,4 @@
  *
  */
 
-int compute_motion(const float *ref, const float *dis, int w, int h, int ref_stride, int dis_stride, double *score);
+int compute_motion(const double *ref, const double *dis, int w, int h, int ref_stride, int dis_stride, double *score);
